+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Localization now scraps `#.#.#-local` workspace dependencies.
 * Add `Var::hold`, `AnyVar::hold_any`.
 * Add `AnyVar::perm`, `VARS::perm`.